week01/backend/requirements.txt
fastapi==0.133.1
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.1
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10

# 测试依赖
pytest==8.3.5
httpx==0.28.1
pytest-asyncio==1.3.0
pytest-cov==6.1.1
